Security

+++ Oodle/Oodle/Controllers/TeachersController.cs
@@ -17,7 +17,6 @@ using System.Text.RegularExpressions;
 
 namespace Oodle.Controllers
 {
-
     [Authorize]
     public class TeachersController : Controller
     {
@@ -27,7 +26,7 @@ namespace Oodle.Controllers
         // GET: Teachers
         private Model1 db = new Model1();
 
-        public ActionResult Index(int classID)
+        public ActionResult test(int classID)
         {
             var idid = User.Identity.GetUserId();
 
@@ -38,6 +37,16 @@ namespace Oodle.Controllers
             {
                 return RedirectToAction("Index", "Class", new { classId = classID });
             }
+            return null;
+        }
+
+
+        public ActionResult Index(int classID)
+        {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
             var teacher = getTVM(classID);
 
             return View("index", "_TeacherLayout", teacher);
@@ -45,6 +54,12 @@ namespace Oodle.Controllers
 
         public ActionResult Accept(int classID, int userID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             db.UserRoleClasses.Where(i => i.UsersID == userID & i.ClassID == classID).ToList().ForEach(x => x.RoleID = 2);
             db.SaveChanges();
 
@@ -59,6 +74,12 @@ namespace Oodle.Controllers
 
         public ActionResult Reject(int classID, int userID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             db.UserRoleClasses.Remove(db.UserRoleClasses.Where(i => i.UsersID == userID & i.ClassID == classID).FirstOrDefault());
             db.SaveChanges();
 
@@ -67,9 +88,13 @@ namespace Oodle.Controllers
 
         public ActionResult Edit(int classID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             ViewBag.id = classID;
-
-
 
             var teacher = getTVM(classID);
 
@@ -97,6 +122,11 @@ namespace Oodle.Controllers
                 AddNotification(notif, classID);
             }
 
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
             db.Classes.Where(i => i.ClassID == classID).ToList().ForEach(x => x.Name = name);
             db.Classes.Where(i => i.ClassID == classID).ToList().ForEach(x => x.Description = desc);
 
@@ -120,6 +150,12 @@ namespace Oodle.Controllers
 
         public ActionResult Delete(int classID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             var list = db.UserRoleClasses.Where(i => i.ClassID == classID);
             Class hasSlack = db.Classes.Where(i => i.ClassID == classID).FirstOrDefault();
             foreach (var i in list)
@@ -194,6 +230,11 @@ namespace Oodle.Controllers
 
             int classID = int.Parse(id);
 
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
 
             var assi = new Assignment();
 
@@ -202,6 +243,7 @@ namespace Oodle.Controllers
             assi.ClassID = classID;
             assi.StartDate = DateTime.Parse(startDate);
             assi.DueDate = DateTime.Parse(dueDate);
+
 
             db.Assignments.Add(assi);
             db.SaveChanges();
@@ -214,6 +256,12 @@ namespace Oodle.Controllers
 
         public ActionResult EditAssignment(int classID, int assignmentID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             var urcL = db.UserRoleClasses.Where(i => i.RoleID == 3 && i.ClassID == classID);
             var list = new List<int>();
 
@@ -244,6 +292,11 @@ namespace Oodle.Controllers
             int assignmentID = int.Parse(assiID);
             int classID = int.Parse(id);
 
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
 
             db.Assignments.Where(i => i.ClassID == classID && i.AssignmentID == assignmentID).ToList().ForEach(x => x.Name = name);
             db.Assignments.Where(i => i.ClassID == classID && i.AssignmentID == assignmentID).ToList().ForEach(x => x.Description = desc);
@@ -273,6 +326,12 @@ namespace Oodle.Controllers
 
         public ActionResult SubmissionView(int classID, int assignmentID)
         {
+            if (test(classID) != null)
+            {
+                return test(classID);
+            }
+
+
             var urcL = db.UserRoleClasses.Where(i => i.RoleID == 3 && i.ClassID == classID);
             var list = new List<int>();
